Add CRD-level marker for top-level desc. field
This commit adds the ability to specify "+kubebuilder:topleveldesc" as a
marker to add "description" as a top-level field in the validation
schema. This is useful for CRDs that would like to have a human-readable
description of the resource. See example below.

```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
description: "L7 policy for getting started using Kubernetes guide"
metadata:
  name: "l7-policy"
spec:
  endpointSelector:
    matchLabels:
      id: app1
  ingress:
  - fromEndpoints:
    - matchLabels:
        id: app2
    toPorts:
    - ports:
      - port: "80"
        protocol: TCP
      rules:
        http:
        - method: "GET"
          path: "/public"
```

Previously with v1beta1 CRDs, any field could be added as long as
`preserveUnknownFields: true`. In order to move CRDs to v1,
`preserveUnknownFields` must be false, meaning that the apiserver will
prune any "unknown" fields before storing it. Therefore, any CRD relying
on having an "unknown" field in its schema will be removed and will fail
validation. In the case of Cilium, it makes heavy use of the top-level
description field to describe what a policy does, for example.

Without this commit, the following error can be observed when applying
the above example (as a v1 CRD):

```
$ kubectl apply -f l7-policy.yaml
error: error validating "test/k8sT/manifests/l7-policy.yaml": error
validating data: ValidationError(CiliumNetworkPolicy): unknown field
"description" in io.cilium.v2.CiliumNetworkPolicy; if you choose to
ignore these errors, turn validation off with --validate=false
```

Hence the need for this commit to allow "description" to be in the CRD
schema.

Signed-off-by: Chris Tarazi <tarazichris@gmail.com>